+++ core/CMakeLists.txt
@@ -20,25 +20,25 @@ message( "# Building core tests: " ${BUILD_CORE_TESTS})
 # Set core source files
 
 set(CUBOS_CORE_SOURCE
-    "src/cubos/debug_render.cpp"
     "src/cubos/log.cpp"
     "src/cubos/settings.cpp"
-
+    
     "src/cubos/io/window.cpp"
     "src/cubos/io/glfw_window.hpp"
     "src/cubos/io/glfw_window.cpp"
 
+    "src/cubos/gl/debug_render.cpp"
     "src/cubos/gl/ogl_render_device.hpp"
     "src/cubos/gl/ogl_render_device.cpp"
 )
 
 set(CUBOS_CORE_INCLUDE
-    "include/cubos/debug_render.hpp"
     "include/cubos/log.hpp"
     "include/cubos/settings.hpp"
     
     "include/cubos/io/window.hpp"
 
+    "include/cubos/gl/debug_render.hpp"
     "include/cubos/gl/render_device.hpp"
 )
